Implemented MLJ Model Interface

+++ src/losses.jl
@@ -10,7 +10,7 @@ Conpute the negative log-likelihood loss for a set of labels `y` under a Gaussia
 - `y`: A 1xn matrix of labels where n is the number of samples.
 """
 function likelihood_loss(μ::Matrix{<:Real}, σ::Matrix{<:Real}, pi::Matrix{<:Real}, y::Matrix{<:Real})
-    return likelihood_loss(Float32.(μ), Float32.(σ), Float32.(pi), Float32.(y))
+    return likelihood_loss(Float64.(μ), Float64.(σ), Float64.(pi), Float64.(y))
 end
 
 function likelihood_loss(μ::Matrix{Float64}, σ::Matrix{Float64}, pi::Matrix{Float64}, y::Matrix{Float64})
